requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
pydantic
cors
python-multipart
fastapi
email-validator